+++ pkg/backend/httpstate/client/client.go
@@ -27,7 +27,6 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/pulumi/pulumi/pkg/apitype"
-	"github.com/pulumi/pulumi/pkg/backend"
 	"github.com/pulumi/pulumi/pkg/diag"
 	"github.com/pulumi/pulumi/pkg/diag/colors"
 	"github.com/pulumi/pulumi/pkg/engine"
@@ -35,6 +34,7 @@ import (
 	"github.com/pulumi/pulumi/pkg/resource/config"
 	"github.com/pulumi/pulumi/pkg/tokens"
 	"github.com/pulumi/pulumi/pkg/util/contract"
+	"github.com/pulumi/pulumi/pkg/util/validation"
 	"github.com/pulumi/pulumi/pkg/workspace"
 )
 
@@ -169,6 +169,11 @@ func (pc *Client) ListStacks(ctx context.Context, projectFilter *string) ([]apit
 	return resp.Stacks, nil
 }
 
+var (
+	// ErrNoPreviousDeployment is returned when there isn't a previous deployment.
+	ErrNoPreviousDeployment = errors.New("no previous deployment")
+)
+
 // GetLatestConfiguration returns the configuration for the latest deployment of a given stack.
 func (pc *Client) GetLatestConfiguration(ctx context.Context, stackID StackIdentifier) (config.Map, error) {
 	latest := struct {
@@ -178,7 +183,7 @@ func (pc *Client) GetLatestConfiguration(ctx context.Context, stackID StackIdent
 	if err := pc.restCall(ctx, "GET", getStackPath(stackID, "updates", "latest"), nil, nil, &latest); err != nil {
 		if restErr, ok := err.(*apitype.ErrorResponse); ok {
 			if restErr.Code == http.StatusNotFound {
-				return nil, backend.ErrNoPreviousDeployment
+				return nil, ErrNoPreviousDeployment
 			}
 		}
 
@@ -214,7 +219,7 @@ func (pc *Client) GetStack(ctx context.Context, stackID StackIdentifier) (apityp
 func (pc *Client) CreateStack(
 	ctx context.Context, stackID StackIdentifier, tags map[apitype.StackTagName]string) (apitype.Stack, error) {
 	// Validate names and tags.
-	if err := backend.ValidateStackProperties(stackID.Stack, tags); err != nil {
+	if err := validation.ValidateStackProperties(stackID.Stack, tags); err != nil {
 		return apitype.Stack{}, errors.Wrap(err, "validating stack properties")
 	}
 
@@ -426,7 +431,7 @@ func (pc *Client) StartUpdate(ctx context.Context, update UpdateIdentifier,
 	tags map[apitype.StackTagName]string) (int, string, error) {
 
 	// Validate names and tags.
-	if err := backend.ValidateStackProperties(update.StackIdentifier.Stack, tags); err != nil {
+	if err := validation.ValidateStackProperties(update.StackIdentifier.Stack, tags); err != nil {
 		return 0, "", errors.Wrap(err, "validating stack properties")
 	}
 
@@ -549,7 +554,7 @@ func (pc *Client) UpdateStackTags(
 	ctx context.Context, stack StackIdentifier, tags map[apitype.StackTagName]string) error {
 
 	// Validate stack tags.
-	if err := backend.ValidateStackTags(tags); err != nil {
+	if err := validation.ValidateStackTags(tags); err != nil {
 		return err
 	}
